+++ src/app/donation/page.tsx
@@ -21,7 +21,7 @@ const MembershipCard = () => {
     }
   };
   return (
-    <div className="mt-20">
+    <div className="">
       <div className="">
         <div className=" p-5">
           <p className="text-center text-4xl font-bold text-gray-700 underline">
